+++ src/argilla/server/policies.py
@@ -88,7 +88,7 @@ class WorkspacePolicy:
 
     @classmethod
     def delete(cls, workspace: Workspace) -> PolicyAction:
-        def is_allowed(actor: User) -> bool:
+        async def is_allowed(actor: User) -> bool:
             return actor.is_owner
 
         return is_allowed
@@ -292,6 +292,15 @@ class DatasetPolicyV1:
         return is_allowed
 
     @classmethod
+    def create_vectors(cls, dataset: Dataset) -> PolicyAction:
+        async def is_allowed(actor: User) -> bool:
+            return actor.is_owner or (
+                actor.is_admin and await _exists_workspace_user_by_user_and_workspace_id(actor, dataset.workspace_id)
+            )
+
+        return is_allowed
+
+    @classmethod
     def delete_records(cls, dataset: Dataset) -> PolicyAction:
         async def is_allowed(actor: User) -> bool:
             return actor.is_owner or (
@@ -433,6 +442,15 @@ class MetadataPropertyPolicyV1:
 
 
 class RecordPolicyV1:
+    @classmethod
+    def get(cls, record: Record) -> PolicyAction:
+        async def is_allowed(actor: User) -> bool:
+            return actor.is_owner or await _exists_workspace_user_by_user_and_workspace_id(
+                actor, record.dataset.workspace_id
+            )
+
+        return is_allowed
+
     @classmethod
     def update(cls, record: Record) -> PolicyAction:
         async def is_allowed(actor: User) -> bool: