+++ src/argilla/server/schemas/v1/datasets.py
@@ -16,12 +16,15 @@ from datetime import datetime
 from typing import Any, Dict, Generic, List, Literal, Optional, TypeVar, Union
 from uuid import UUID
 
+from fastapi import HTTPException, Query
+from pydantic import BaseModel, PositiveInt, conlist, constr, root_validator, validator
 from fastapi import Query
 from pydantic import BaseModel, PositiveInt, conint, conlist, constr, root_validator, validator
 from pydantic import Field as PydanticField
 from pydantic.generics import GenericModel
 from pydantic.utils import GetterDict
 
+from argilla.server.enums import RecordInclude
 from argilla.server.schemas.base import UpdateSchema
 from argilla.server.schemas.v1.records import RecordUpdate
 from argilla.server.schemas.v1.suggestions import Suggestion, SuggestionCreate
@@ -373,6 +376,11 @@ class VectorSettings(BaseModel):
     class Config:
         orm_mode = True
 
+    def check_vector(self, value: List[float]) -> None:
+        num_elements = len(value)
+        if num_elements != self.dimensions:
+            raise ValueError(f"vector must have {self.dimensions} elements, got {num_elements} elements")
+
 
 class VectorsSettings(BaseModel):
     items: List[VectorSettings]
@@ -471,6 +479,7 @@ class RecordCreate(BaseModel):
     external_id: Optional[str]
     responses: Optional[List[UserResponseCreate]]
     suggestions: Optional[List[SuggestionCreate]]
+    vectors: Optional[Dict[str, List[float]]]
 
     @validator("responses")
     def check_user_id_is_unique(cls, values: Optional[List[UserResponseCreate]]) -> Optional[List[UserResponseCreate]]:
@@ -480,16 +489,14 @@ class RecordCreate(BaseModel):
         user_ids = []
         for value in values:
             if value.user_id in user_ids:
-                raise ValueError(f"Responses contains several responses for the same user_id: {str(value.user_id)!r}")
+                raise ValueError(f"'responses' contains several responses for the same user_id={str(value.user_id)!r}")
             user_ids.append(value.user_id)
 
         return values
 
 
 class RecordsCreate(BaseModel):
-    items: List[RecordCreate] = PydanticField(
-        ..., min_items=RECORDS_CREATE_MIN_ITEMS, max_items=RECORDS_CREATE_MAX_ITEMS
-    )
+    items: conlist(item_type=RecordCreate, min_items=RECORDS_CREATE_MIN_ITEMS, max_items=RECORDS_CREATE_MAX_ITEMS)
 
 
 class RecordUpdateWithId(RecordUpdate):
@@ -497,6 +504,7 @@ class RecordUpdateWithId(RecordUpdate):
 
 
 class RecordsUpdate(BaseModel):
+    # TODO: review this definition and align to create model
     items: List[RecordUpdateWithId] = PydanticField(
         ..., min_items=RECORDS_UPDATE_MIN_ITEMS, max_items=RECORDS_UPDATE_MAX_ITEMS
     )